Move components to the correct folder
