+++ src/components/Buttons/RemoveOperatorButton/index.js
