+++ src/components/Buttons/index.js
@@ -1,12 +1,17 @@
-export { default as CompareResultsButton } from './CompareResultsButton'
-export { default as DataViewButton } from './DataViewButton'
-export { default as DebugButton } from './DebugButton'
-export { default as DeleteButton } from './DeleteButton'
-export { default as ErrorButton } from './ErrorButton'
-export { default as InfoButton } from './InfoButton'
-export { default as LogsButton } from './LogsButton'
-export { default as PrepareDeploymentsButton } from './PrepareDeploymentsButton'
-export { default as RunDeploymentButton } from './RunDeploymentButton'
-export { default as SaveTemplateButton } from './SaveTemplateButton'
-export { default as UploadButton } from './UploadButton'
-export { default as UsingDeploymentsButton } from './UsingDeploymentsButton'
+export { default as CompareResultsButton } from './CompareResultsButton';
+export { default as DataViewButton } from './DataViewButton';
+export { default as DebugButton } from './DebugButton';
+export { default as DeleteButton } from './DeleteButton';
+export { default as ErrorButton } from './ErrorButton';
+export { default as InfoButton } from './InfoButton';
+export { default as InterruptTrainExperimentButton } from './InterruptTrainExperimentButton';
+export { default as LogsButton } from './LogsButton';
+export { default as PrepareDeploymentsButton } from './PrepareDeploymentsButton';
+export { default as RemoveOperatorButton } from './RemoveOperatorButton';
+export { default as ResultsButtonBar } from './ResultsButtonBar';
+export { default as RunDeploymentButton } from './RunDeploymentButton';
+export { default as SaveTemplateButton } from './SaveTemplateButton';
+export { default as TrainExperimentButton } from './TrainExperimentButton';
+export { default as UploadButton } from './UploadButton';
+export { default as UploadInferenceTestButton } from './UploadInferenceTestButton';
+export { default as UsingDeploymentsButton } from './UsingDeploymentsButton';
