+++ src/components/Content/ExperimentsContent/Experiment/Drawer/ResultsDrawer/PlotResult/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 //COMPONENTS
-import ImageLightbox from './ImageLightbox';
+import ImageLightbox from 'components/ImageLightbox';
 
 /**
  * Plot Result.
