+++ src/components/Content/ExperimentsContent/_/FlowDrop/index.js
@@ -1,21 +1,21 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { useParams } from "react-router-dom";
-import { ReactFlowProvider } from "react-flow-renderer";
+import React from 'react';
+import PropTypes from 'prop-types';
+import { useParams } from 'react-router-dom';
+import { ReactFlowProvider } from 'react-flow-renderer';
 
 //COMPONENTS
-import ExperimentEmpty from "../../Experiment/ExperimentEmpty";
-import ExperimentHeader from "../../Experiment/ExperimentHeader/_/Container";
-import ExperimentFlow from "../../Experiment/ExperimentFlow/_/Container";
-import CustomDragLayer from "../CustomDragLayer";
+import ExperimentEmpty from 'components/ExperimentEmpty';
+import ExperimentHeader from '../../Experiment/ExperimentHeader/_/Container';
+import ExperimentFlow from '../../Experiment/ExperimentFlow/_/Container';
+import CustomDragLayer from '../CustomDragLayer';
 
-import "./style.less";
+import './style.less';
 
 const FlowDrop = () => {
   const { experimentId } = useParams();
 
   return (
-    <div className="custom-flow">
+    <div className='custom-flow'>
       <ReactFlowProvider>
         <ExperimentHeader />
         {experimentId ? <ExperimentFlow /> : <ExperimentEmpty />}
