+++ src/components/ImageLightbox/styles.less
