+++ src/containers/EditTitleContainer/index.js
@@ -1,0 +1,3 @@
+import EditTitleContainer from './EditTitleContainer';
+
+export default EditTitleContainer;
